requirements.txt
ffmpeg-python==0.2.0
matplotlib==3.9.2
numpy==1.26.4
opencv-python==4.9.0
pytest==8.3.5
scipy==1.14.1
